refactor(storybook): refactor component interfaces
