+++ components/common/Button/Button.interface.ts
@@ -1,0 +1,8 @@
+export interface StyledButtonProps {
+  isRounded?: boolean;
+  type?: 'button' | 'submit' | 'reset';
+}
+
+export interface ButtonProps extends StyledButtonProps {
+  text: string;
+}
